Changed climb solenoid to double, added button fnction and climb command
Climb solenoid was changed to a double using boolean logic.  It needs to
be changed to WPI method for double just for good measure.  Also, we
have changed comment convention.  We will now use /* to give
instructions/notes and // is reserved for commenting out code.  Also,
button naming convention has change to ds for drivestick and ss for
shooter stick, followed by capital button label, ie dsA for button A on
the drivestick.